Move `MustDefaultConfigWithRandomPorts` and `TCPRandomPort` to `testutils` package
Having the ability to spawn an in-process OpenFGA for testing purposes
is very handy and has been very useful for us.

A recent PR [1] moved it to an internal package, which made it hard to
use. So, this moves the aforementioned functions to an exported package
so folks can keep benefiting from these handy functions.

[1] https://github.com/openfga/openfga/pull/1468
