+++ tests/check/check_test.go
@@ -55,7 +55,7 @@ func TestCheckLogs(t *testing.T) {
 	// defer goleak.VerifyNone(t)
 
 	// create mock OTLP server
-	otlpServerPort, otlpServerPortReleaser := config.TCPRandomPort()
+	otlpServerPort, otlpServerPortReleaser := testutils.TCPRandomPort()
 	localOTLPServerURL := fmt.Sprintf("localhost:%d", otlpServerPort)
 	otlpServerPortReleaser()
 	_ = mocks.NewMockTracingServer(t, otlpServerPort)
